+++ model_zoo/official/cv/faster_rcnn/src/FasterRcnn/fpn_neck.py
@@ -26,12 +26,12 @@ context.set_context(mode=context.GRAPH_MODE, device_target="Ascend")
 
 def bias_init_zeros(shape):
     """Bias init method."""
-    return Tensor(np.array(np.zeros(shape).astype(np.float32)).astype(np.float16))
+    return Tensor(np.array(np.zeros(shape).astype(np.float32)).astype(np.float32))
 
 def _conv(in_channels, out_channels, kernel_size=3, stride=1, padding=0, pad_mode='pad'):
     """Conv2D wrapper."""
     shape = (out_channels, in_channels, kernel_size, kernel_size)
-    weights = initializer("XavierUniform", shape=shape, dtype=mstype.float16).to_tensor()
+    weights = initializer("XavierUniform", shape=shape, dtype=mstype.float32).to_tensor()
     shape_bias = (out_channels,)
     biass = bias_init_zeros(shape_bias)
     return nn.Conv2d(in_channels, out_channels,
