+++ model_zoo/official/cv/faster_rcnn/src/FasterRcnn/proposal_generator.py
@@ -106,7 +106,7 @@ class Proposal(nn.Cell):
         self.tile = P.Tile()
         self.set_train_local(config, training=True)
 
-        self.multi_10 = Tensor(10.0, mstype.float16)
+        self.multi_10 = Tensor(10.0, mstype.float32)
 
     def set_train_local(self, config, training=True):
         """Set training flag."""
@@ -133,7 +133,7 @@ class Proposal(nn.Cell):
         self.topKv2 = P.TopK(sorted=True)
         self.topK_shape_stage2 = (self.max_num, 1)
         self.min_float_num = -65536.0
-        self.topK_mask = Tensor(self.min_float_num * np.ones(total_max_topk_input, np.float16))
+        self.topK_mask = Tensor(self.min_float_num * np.ones(total_max_topk_input, np.float32))
 
     def construct(self, rpn_cls_score_total, rpn_bbox_pred_total, anchor_list):
         proposals_tuple = ()
@@ -164,16 +164,16 @@ class Proposal(nn.Cell):
 
             rpn_cls_score = self.reshape(rpn_cls_score, self.reshape_shape)
             rpn_cls_score = self.activation(rpn_cls_score)
-            rpn_cls_score_process = self.cast(self.squeeze(rpn_cls_score[::, 0::]), mstype.float16)
+            rpn_cls_score_process = self.cast(self.squeeze(rpn_cls_score[::, 0::]), mstype.float32)
 
-            rpn_bbox_pred_process = self.cast(self.reshape(rpn_bbox_pred, (-1, 4)), mstype.float16)
+            rpn_bbox_pred_process = self.cast(self.reshape(rpn_bbox_pred, (-1, 4)), mstype.float32)
 
             scores_sorted, topk_inds = self.topKv2(rpn_cls_score_process, self.topK_stage1[idx])
 
             topk_inds = self.reshape(topk_inds, self.topK_shape[idx])
 
             bboxes_sorted = self.gatherND(rpn_bbox_pred_process, topk_inds)
-            anchors_sorted = self.cast(self.gatherND(anchors, topk_inds), mstype.float16)
+            anchors_sorted = self.cast(self.gatherND(anchors, topk_inds), mstype.float32)
 
             proposals_decode = self.decode(anchors_sorted, bboxes_sorted)
 
@@ -188,7 +188,7 @@ class Proposal(nn.Cell):
 
         _, _, _, _, scores = self.split(proposals)
         scores = self.squeeze(scores)
-        topk_mask = self.cast(self.topK_mask, mstype.float16)
+        topk_mask = self.cast(self.topK_mask, mstype.float32)
         scores_using = self.select(masks, scores, topk_mask)
 
         _, topk_inds = self.topKv2(scores_using, self.max_num)
